releass version

+++ app/src/main/java/com/tikonsil/tikonsil509/domain/model/SendRechargeResponse.kt
@@ -1,7 +1,7 @@
 package com.tikonsil.tikonsil509.domain.model
 
 data class SendRechargeResponse(
-    val balance: Int,
+    val balance: Float,
     val destination: String,
     val error_code: Int,
     val recharge_id: Int,
